Revert "Develop se corrige bug de issues 46 y 48"
